Convert Field component from .js to .tsx
